+++ kaithem/src/chandler/groups.py
@@ -1824,6 +1824,7 @@ class Group:
                 for i in dt:
                     i[1] = tagpoints.normalize_tag_name(i[1])
                     # Upgrade legacy format
+                    # TODO remove legacy stuff
                     if len(i) == 2:
                         i.append({"type": "null"})